Changed the order of appspec hooks and added validation in stop container

+++ deployment/scripts/stop_container.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
-#Stop Container aws-devops
-docker ps -aq --filter "name=aws-devops" | xargs docker stop
-#Remove container aws-devops
-docker ps -aq --filter "name=aws-devops" | xargs docker rm
+running_app=`docker ps -aq | wc -l`
+
+if [[ $running_app -eq 1 ]]; then
+    #Stop Container aws-devops
+    docker ps -aq --filter "name=aws-devops" | xargs docker stop
+    #Remove container aws-devops
+    docker ps -aq --filter "name=aws-devops" | xargs docker rm
+else echo "There are no running containers"
